xiugai

+++ flappyim/Classes/Public/ChatSession.h
@@ -31,7 +31,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @property(nonatomic,copy) NSString* sessionOffset;
 
-@property(nonatomic,assign) NSInteger sessionStamp;
+@property(nonatomic,assign) long sessionStamp;
 
 @property(nonatomic,copy) NSString* sessionCreateDate;
